+++ release-notes.md
@@ -1,5 +1,7 @@
 # Release Notes
 
+- Allow library specification in template lookup
+
 ## 2.2.1 - June 11th 2013
 
 - Hack around bower config issue when passing config files on CLI